requirements.txt
pytesseract==0.3.7
opencv_python==4.5.2.52
numpy==1.20.3
Pillow==8.2.0
